+++ tidy3d/components/types.py
@@ -180,6 +180,7 @@ ObsGridArray = Union[Tuple[float, ...], ArrayLike[float, 1]]
 
 Ax = Axes
 PlotVal = Literal["real", "imag", "abs"]
+FieldVal = Literal["real", "imag", "abs", "abs^2", "dB"]
 
 """ mode tracking """